Update initialization of transforms in state
